Refactoring of the event system + Changed declarations of all primitive types using their alias to be more inline with the C# design guidelines + long types in DTOs

+++ Client/Unity Project/Assets/Scripts/Scenes/Game/AttackBatchingCommandInterceptor.cs
@@ -16,7 +16,7 @@ namespace Assets.Scripts.Scenes.Game
         public float BatchSendInterval = 0.5f;
 
         private WebSocketService webSocketService;
-        private int postponedAttackValue = 0;
+        private long postponedAttackValue = 0;
         private float lastBatchSendTime = 0;
 
         void Start()
